+++ app/src/main/java/com/example/bankapp/data/repository/TransactionActionRepositoryImpl.kt
@@ -1,23 +1,33 @@
 package com.example.bankapp.data.repository
 
 import com.example.bankapp.core.ResultState
+import com.example.bankapp.data.remote.model.TransactionDto
 import com.example.bankapp.data.remote.model.UpdateTransactionRequest
 import com.example.bankapp.di.Local
 import com.example.bankapp.di.NetworkChecker
 import com.example.bankapp.di.Remote
+import com.example.bankapp.domain.model.Transaction
 import com.example.bankapp.domain.model.TransactionEdit
 import com.example.bankapp.domain.repository.TransactionActionRepository
+import com.example.bankapp.domain.repository.WriteRepository
 import javax.inject.Inject
 
 class TransactionActionRepositoryImpl @Inject constructor(
     @Remote private val remoteTransactionActionRepositoryImpl: TransactionActionRepository,
     @Local private val localTransactionActionRepositoryImpl: TransactionActionRepository,
+    private val writeCategoryRepositoryImpl: WriteRepository<TransactionDto>,
     private val networkChecker: NetworkChecker
 ) : TransactionActionRepository {
 
-    override suspend fun addTransaction(request: UpdateTransactionRequest): ResultState<Unit> {
+    override suspend fun addTransaction(request: UpdateTransactionRequest): ResultState<TransactionDto?> {
         return if (networkChecker.isOnline()) {
-            remoteTransactionActionRepositoryImpl.addTransaction(request)
+
+            val result = remoteTransactionActionRepositoryImpl.addTransaction(request)
+            if (result is ResultState.Success) {
+
+                writeCategoryRepositoryImpl.addDb(entity = result.data!!)
+            }
+            ResultState.Success(null)
         } else {
             localTransactionActionRepositoryImpl.addTransaction(request)
         }
@@ -26,14 +36,23 @@ class TransactionActionRepositoryImpl @Inject constructor(
     override suspend fun editTransaction(
         transactionId: Int,
         request: UpdateTransactionRequest
-    ): ResultState<Unit> {
-        return if (networkChecker.isOnline()) {
-            remoteTransactionActionRepositoryImpl.editTransaction(
+    ): ResultState<Transaction?> {
+        if (networkChecker.isOnline()) {
+            val result = remoteTransactionActionRepositoryImpl.editTransaction(
                 transactionId = transactionId,
                 request = request
             )
+            if (result is ResultState.Success) {
+                localTransactionActionRepositoryImpl.editTransaction( //тут по идее можем редактировать транзу которой нет в бд или нет?
+                    transactionId = transactionId,
+                    request = request
+                )
+
+            }
+            return ResultState.Success(null)
+
         } else {
-            localTransactionActionRepositoryImpl.editTransaction(
+            return localTransactionActionRepositoryImpl.editTransaction(
                 transactionId = transactionId,
                 request = request
             )
@@ -41,19 +60,24 @@ class TransactionActionRepositoryImpl @Inject constructor(
     }
 
     override suspend fun getTransactionById(transactionId: Int): ResultState<TransactionEdit> {
-        return  if(networkChecker.isOnline()) {
-            remoteTransactionActionRepositoryImpl.getTransactionById(transactionId=transactionId)
+        return if (networkChecker.isOnline()) {
+            remoteTransactionActionRepositoryImpl.getTransactionById(transactionId = transactionId)
         } else {
-            localTransactionActionRepositoryImpl.getTransactionById(transactionId=transactionId)
+            localTransactionActionRepositoryImpl.getTransactionById(transactionId = transactionId)
         }
     }
 
     override suspend fun deleteTransactionById(transactionId: Int): ResultState<Unit> {
         if (networkChecker.isOnline()) {
-            val result = remoteTransactionActionRepositoryImpl.deleteTransactionById(transactionId=transactionId)
+            val result =
+                remoteTransactionActionRepositoryImpl.deleteTransactionById(transactionId = transactionId)
+
+            if (result is ResultState.Success) {
+                localTransactionActionRepositoryImpl.deleteTransactionById(transactionId = transactionId)
+            }
             return result
         } else {
-            return localTransactionActionRepositoryImpl.deleteTransactionById(transactionId=transactionId)
+            return localTransactionActionRepositoryImpl.deleteTransactionById(transactionId = transactionId)
         }
     }
 }
